Move first chunk of structs into specifc headers
