+++ src/OpenLoco/src/GameCommands/Cheats/Cheat.h
@@ -4,6 +4,41 @@
 
 namespace OpenLoco::GameCommands
 {
+    struct VehicleApplyShuntCheatArgs
+    {
+        static constexpr auto command = GameCommand::vehicleApplyShuntCheat;
+
+        VehicleApplyShuntCheatArgs() = default;
+        explicit VehicleApplyShuntCheatArgs(const registers& regs)
+            : head(EntityId(regs.cx))
+        {
+        }
+
+        EntityId head;
+
+        explicit operator registers() const
+        {
+            registers regs;
+            regs.cx = enumValue(head);
+            return regs;
+        }
+    };
+
+    struct ApplyFreeCashCheatArgs
+    {
+        static constexpr auto command = GameCommand::applyFreeCashCheat;
+
+        ApplyFreeCashCheatArgs() = default;
+        explicit ApplyFreeCashCheatArgs(const registers&)
+        {
+        }
+
+        explicit operator registers() const
+        {
+            return registers();
+        }
+    };
+
     enum class CheatCommand : uint8_t
     {
         acquireAssets,
