+++ src/OpenLoco/src/GameCommands/Company/RenameCompanyName.cpp
@@ -1,3 +1,4 @@
+#include "RenameCompanyName.h"
 #include "Economy/Expenditures.h"
 #include "Engine/Limits.h"
 #include "GameCommands/GameCommands.h"
