+++ src/OpenLoco/src/Windows/CompanyWindow.cpp
@@ -6,6 +6,9 @@
 #include "GameCommands/Company/BuildCompanyHeadquarters.h"
 #include "GameCommands/Company/ChangeCompanyColour.h"
 #include "GameCommands/Company/ChangeLoan.h"
+#include "GameCommands/Company/RemoveCompanyHeadquarters.h"
+#include "GameCommands/Company/RenameCompanyName.h"
+#include "GameCommands/Company/RenameCompanyOwner.h"
 #include "GameCommands/GameCommands.h"
 #include "GameState.h"
 #include "Graphics/ImageIds.h"
