+++ src/OpenLoco/src/World/CompanyManager.cpp
@@ -5,6 +5,9 @@
 #include "Economy/Economy.h"
 #include "Entities/EntityManager.h"
 #include "Entities/Misc.h"
+#include "GameCommands/Company/RenameCompanyName.h"
+#include "GameCommands/Company/RenameCompanyOwner.h"
+#include "GameCommands/Company/UpdateOwnerStatus.h"
 #include "GameCommands/GameCommands.h"
 #include "GameState.h"
 #include "GameStateFlags.h"
